+++ backend/wb_tun_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package backend
+
+func prepareWBTun() error {
+	return extractWintun()
+}